fix _get_relay_op
