+++ python/tvm/relay/frontend/onnx.py
@@ -9,21 +9,10 @@ from tvm import relay
 from .. import ir_pass
 from .. import expr as _expr
 from .. import op as _op
+from .common import _get_relay_op
 from .common import AttrCvt
 
 __all__ = ['from_onnx']
-
-def _get_relay_op(op_name):
-    try:
-        op = getattr(_op, op_name)
-    except AttributeError:
-        try:
-            op = getattr(_op.nn, op_name)
-        except AttributeError:
-            op = getattr(_op.image, op_name)
-    if not op:
-        raise RuntimeError("Unable to map op_name {} to relay".format(op_name))
-    return op
 
 def dimension_picker(prefix, surfix=''):
     def _impl(attr):
